Added SetRendermesh(); moved TryGetRendermesh in Query. Correction in IRenderMesh().
