+++ TriangleNet_Engine/TriangleNet_Engine.csproj
@@ -100,6 +100,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Query\Representation\RenderMesh\TryGetRendermesh.cs" />
+    <Compile Include="Modify\Representation\RenderMesh\SetRendermesh.cs" />
     <Compile Include="Compute\DelaunayTriangulation.cs" />
     <Compile Include="Compute\Representation\GeometricalRepresentation\Element0D\Point.cs" />
     <Compile Include="Compute\Representation\GeometricalRepresentation\Element0D\Structure\Node.cs" />
